Initial domain registry connector prototype.
The domain registry uses a possible new possible architecture for
a modular "connector" and/or "protostubs" system.

In this architecture each connector is a vertx worker verticle (written
in javascript) that communicate with the main verticle using the
EventBus.

In the current prototype there is a RegistryConnector IComponent that
when receives registry related messages sends them to the connector for
processing.

TODO:
- Better connector messages definition;
- Dynamically load all the connectors/verticles.

Issue #2